Prepares project. Adds stubs to rrequire.
This sets up a Graph to be ready to do code analysis.  Adding the stubs
ensures stubs get added in to the analysis.  The stubs, kept in the
rsense-core repository, are the foundation Rsense builds on.  By having
a very accurate representation of the types Ruby has in its core-lib and
some of its stdlib, Rsense can make better guesses about unknown code.