Performance improvements and bug fix
Dependencies are now mostly loaded in a background thread, as loading
the entire dependency graph was too slow.  Now an initial set of
dependencies is loaded during the init, and then further branches in
the graph are switched to loading in a background thread.

Further, some variables needed to be reset between each request in
order to ensure that the completion results didn't just grow. This is
due to the way the constraint graph selects matches.

Finally, added `require_relative` which dramatically improved results ,
along with some Opal macros making Opal code work a little better. In
the future, it should be possible to add special methods for handling
rspec and minitest/spec `describe` and `it` blocks.  We could
potentially allow a plug-in system, even, so that plugins could help
with handling other places where metaprogramming has made type-inference
fail.