Sets parser to use comments.
Rsense uses annotations via comments in the stub files
that it builds up its internal representation of ruby with to
resolve ambiguities of type.

These had been turned off in the parser configuration, causing
blocks, procs, etc to not work. Because the tests for those
were using a separate instantiation of the parser, the tests
had been passing.

Now we specifically test for comments from the parser Rsense
uses itself.